+++ aggligator-transport-tcp/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Aggligator TCP transport will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.0 - 2026-08-26
+### Changed
+- **breaking:** `tcp_server` and `tls_server` have been replaced by `tcp_listen` and
+  `tls_listen`, which return a `Listener` providing an `accept` method, instead of
+  taking a work function that is spawned for each incoming connection
+
 ## 0.5.0 - 2026-08-13
 ### Changed
 - update aggligator to 0.12.0